Change time columns to double for increased precision in MySQL

+++ database/migrations/2019_03_19_185331_create_checkresults_table.php
@@ -18,7 +18,7 @@ class CreateCheckresultsTable extends Migration
                 $table->increments('id');
                 $table->integer('resultfile_id')->unsigned()->nullable();
                 $table->string('sessionid');
-                $table->float('checktime');
+                $table->double('checktime');
                 $table->integer('checkmemory');
                 $table->timestamps();
